Further fix

+++ src/models/piranha_arguments.rs
@@ -108,7 +108,7 @@ pub struct PiranhaArguments {
   #[get = "pub"]
   #[builder(default = "default_cleanup_comments_buffer()")]
   #[clap(long, default_value_t = default_cleanup_comments_buffer())]
-  cleanup_comments_buffer: usize,
+  cleanup_comments_buffer: i32,
 
   /// Enables deletion of associated comments
   #[get = "pub"]
@@ -165,7 +165,7 @@ impl PiranhaArguments {
   fn py_new(
     language: String, substitutions: &PyDict, path_to_configurations: Option<String>,
     rule_graph: Option<RuleGraph>, path_to_codebase: Option<String>, code_snippet: Option<String>,
-    dry_run: Option<bool>, cleanup_comments: Option<bool>, cleanup_comments_buffer: Option<usize>,
+    dry_run: Option<bool>, cleanup_comments: Option<bool>, cleanup_comments_buffer: Option<i32>,
     number_of_ancestors_in_parent_scope: Option<u8>, delete_consecutive_new_lines: Option<bool>,
     global_tag_prefix: Option<String>, delete_file_if_empty: Option<bool>,
     path_to_output_summary: Option<String>,
